[fix] Cleanup Damage Calculation to Improve Hit Rate/Damage Amounts (#65)
closes https://github.com/TimPoliquin/unreal-electric-castle/issues/64

- Add ability damage magnitude. Abilities no longer have explicit damage
but instead provide a magnitude that can modify damage based on
attributes.
- Major changes to hit chance, evade chance, and armor damage mitigation
- Removed DurationDamage effect. Periodic damage is applied as a
combination of timers and instant damage instead. This is because the
damage context is shared for the duration of the ability, resulting in
bleeding of effects from previous executions.
- Removed "block" concept from damage calculation. Blocking will have to
be an explicit ability, and will prevent the damage effect from being
applied.
- Added "evade" concept
- Updated damage text widget to streamline implementation for normal
hits, critical hits, and misses.